// src/components/cdk/cdk-config-notification/models/cdk-config-notification.model.ts
/**Literales*/
export interface ICDKConfigLiterals {
	titleConfi: string;
	subTitle: string;
	subText: string;
	boton: string;
	botonSave: string;
	textPlatform: string;
	textSubPlatform: string;
}

export interface ICDKConfigCheckGeneral {
	disabled: boolean;
	indeterminate: boolean;
	checkAll: boolean;
}

export interface ICDKConfigData {
	img?: string;
	nameApp: string;
	descriptionApp: string;
	checked: boolean;
	imgEnabled: boolean;
	checkisDisabled: boolean;
	idCheck: string;
}